sql: check number of arguments during parsing
Prior to this patch, the ability to accept a given number of arguments
for a function with a variable number of arguments was checked at
runtime. After this patch, it will be checked during parsing. For
functions with the constant number of arguments, it is always checked
during parsing.

Part of #6105