sql: modify signature of TRIM()
Currently, to determine type of the result of SQL built-in function
TRIM() we have to know type of the last argument. This patch changes
the way that function arguments are passed, so after this patch, the
type of the result can be determined by knowing the type of the first
argument.

Additionally, these changes fix a bug when using TRIM with the keywords
BOTH, LEADING, or TRAILING will lead to losing collation.

Needed for #6105
Closes #6299