sql: define ephemeral space fields
Prior to this patch, most ephemeral space fields were defined using the
SCALAR type. After this patch, almost all fields will be properly
defined. However, there are still cases where SCALAR will be set by
default. For example, in IN, where rules is still not defined (#4692).
Another example is when a field is not resolved because of a too complex
query.

Part of #6213